requirements.txt
torch
torchvision
termcolor
pillow
tqdm